refactor register/auth/nouvelle classe emergency

+++ MQ_diving_logs/viewsets/diving_log_viewset.py
@@ -9,7 +9,9 @@ class DivingLogViewSet(viewsets.ModelViewSet):
     serializer_class = DivingLogSerializer
 
     def create(self, request, *args, **kwargs):
-        """Customize the creation of a diving log."""
+        # Mettre le statut par défaut à EN_ATTENTE lors de la création
+        request.data['status'] = 'EN_ATTENTE'
+
         serializer = self.get_serializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -17,28 +19,36 @@ class DivingLogViewSet(viewsets.ModelViewSet):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def update(self, request, *args, **kwargs):
-        """Customize the update of a diving log."""
-        instance = self.get_object()
-        serializer = self.get_serializer(instance, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        user = request.user
+
+        if hasattr(user, 'role') and user.role == 'FORMATEUR':
+            instance = self.get_object()
+            serializer = self.get_serializer(instance, data=request.data, partial=True)
+
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
+
+        # Si ce n'est pas un formateur ou si la validation échoue, refuser
+        return Response(status=status.HTTP_403_FORBIDDEN, data={"message": "Permission refusée."})
 
     def destroy(self, request, *args, **kwargs):
-        """Customize the deletion of a diving log."""
-        instance = self.get_object()
-        instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        user = request.user
+
+        if hasattr(user, 'role') and user.role == 'FORMATEUR':
+            instance = self.get_object()
+            instance.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+        # Si ce n'est pas un formateur, refuser
+        return Response(status=status.HTTP_403_FORBIDDEN, data={"message": "Permission refusée."})
 
     def list(self, request, *args, **kwargs):
-        """Customize the listing of diving logs."""
         queryset = self.filter_queryset(self.get_queryset())
         serializer = self.get_serializer(queryset, many=True)
         return Response(serializer.data)
 
     def retrieve(self, request, *args, **kwargs):
-        """Customize the retrieval of a specific diving log."""
         instance = self.get_object()
         serializer = self.get_serializer(instance)
         return Response(serializer.data)
